feat(breakdown-tag): add support for expression comments
